services/openclaw-bridge/requirements.txt
fastapi
uvicorn[standard]
websockets
paho-mqtt>=2.0
loguru
python-dotenv
aioimaplib
aiohttp
